[development-plastic-transaction] feat: add points transfer and listing to users

+++ feature/transaction/plastic/src/main/java/com/clastic/transaction/plastic/PlasticTransactionViewModel.kt
@@ -128,7 +128,7 @@ internal class PlasticTransactionViewModel @Inject constructor(
                 date = _currentDate.value,
                 dropPointId = _dropPoint.value.id,
                 ownerId = userRepository.getLoggedInUser()?.email ?: "",
-                totalPoints = _points.value,
+                totalPoints = _points.value.toLong(),
                 userId = _user.value.email,
                 plasticTransactionItemList = _plasticTransactionItemList.value,
                 onPostSuccess = {
